Download orgs on load
On load page, we download all orgs that the user is a part of. Also, if the user is not authenticated, we do not query for events or orgs. Created a store for orgs, removed unused imports, and refactored types so that the returned query type is found in types.ts